fix(web): fix prefixClassnames feat
