+++ apps/web/tsconfig.json
@@ -18,7 +18,8 @@
     "paths": {
       "@modules/*": ["modules/*"],
       "@components/*": ["components/*"],
-      "@lib/*": ["lib/*"]
+      "@lib/*": ["lib/*"],
+      "@utils/*": ["utils/*"]
     }
   },
   "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx"],
